# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevLog+",
    "description": "Single-user developer journal for technical learning and skill maintenance.\n\nDevLog+ has two engines:\n\n- **Learning engine** — builds a Knowledge Profile from journal entries, quizzes, and feedback. Understands what you know, where you're weak, and what to explore next.\n- **Practice engine** — generates weekly Go micro-projects calibrated to your practical level.\n\n## Core workflow\n\n1. **Journal** — write technical entries (typed or dictated) about what you learn.\n2. **Profile** — the system builds and maintains a Knowledge Profile nightly.\n3. **Quiz** — weekly free-text quizzes probe your understanding.\n4. **Readings** — curated reading recommendations from your allowlisted domains.\n5. **Projects** — weekly Go micro-projects with bugs, features, and refactors.\n6. **Feedback** — thumbs-up/down reactions and feedforward notes steer future content.\n7. **Triage** — items the system can't resolve are surfaced for your clarification.\n",
    "license": {
      "name": "Private — single-user local application"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quizzes/sessions": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "List quiz sessions",
        "description": "Return quiz sessions ordered by creation date (most recent first).\n\nEach session represents a weekly quiz with a configurable number of\nfree-text questions generated from the Knowledge Profile.",
        "operationId": "list_sessions_api_v1_quizzes_sessions_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of sessions to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of sessions to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum sessions to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum sessions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of quiz sessions, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizSessionResponse"
                  },
                  "title": "Response List Sessions Api V1 Quizzes Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quizzes/sessions/{session_id}": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "Get quiz session details",
        "description": "Retrieve a specific quiz session with its questions, submitted answers,\nand LLM-generated evaluations.\n\nEvaluations include correctness (full / partial / incorrect), depth\nassessment, explanation, and confidence score.",
        "operationId": "get_session_api_v1_quizzes_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz session with all questions, answers, and evaluations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSessionDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quiz session not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quizzes/questions/{question_id}/answer": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Submit quiz answer",
        "description": "Submit a free-text answer to a quiz question.\n\nOnly free-text answers are accepted (no multiple choice).  Each question\ncan only be answered once.  The answer will be evaluated by an LLM judge\nwhen the session is completed.",
        "operationId": "submit_answer_api_v1_quizzes_questions__question_id__answer_post",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAnswerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The recorded answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAnswerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quiz question not found"
          },
          "422": {
            "description": "Validation error — answer_text is required"
          }
        }
      }
    },
    "/api/v1/quizzes/sessions/{session_id}/complete": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Complete quiz session",
        "description": "Mark a quiz session as completed.\n\nAll answers should be submitted before completing.  Once completed, the\nquiz evaluation pipeline will assess each answer and update the Knowledge\nProfile accordingly.",
        "operationId": "complete_session_api_v1_quizzes_sessions__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The completed quiz session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quiz session not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List weekly projects",
        "description": "Return weekly projects ordered by issue date (most recent first).\n\nEach project is a self-contained Go codebase with source files, tests,\nand a set of tasks (bugs, features, refactors, optimizations).",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of projects to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of projects to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum projects to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum projects to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of weekly Go projects, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyProjectResponse"
                  },
                  "title": "Response List Projects Api V1 Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get project details",
        "description": "Retrieve a specific weekly project with its tasks and evaluation.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project with tasks, metadata, and evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyProjectDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/submit": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Submit project for evaluation",
        "description": "Submit a completed project for AI evaluation.\n\nThe project evaluation pipeline will assess code quality, task completion,\nand test results.  The evaluation must complete before the next weekly\nproject is issued.",
        "operationId": "submit_project_api_v1_projects__project_id__submit_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The project with updated status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyProjectResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found or not in a submittable state"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "issued",
          "in_progress",
          "submitted",
          "evaluated",
          "skipped"
        ],
        "title": "ProjectStatus"
      },
      "ProjectSubmitRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional notes about the submission (e.g. challenges faced, shortcuts taken)",
            "examples": [
              "Focused on the refactor task first; ran out of time on the optimization."
            ]
          }
        },
        "type": "object",
        "title": "ProjectSubmitRequest",
        "description": "Submit a completed project for AI evaluation."
      },
...
      "QuizAnswerCreate": {
        "properties": {
          "answer_text": {
            "type": "string",
            "minLength": 1,
            "title": "Answer Text",
            "description": "The user's free-text answer (no multiple choice)",
            "examples": [
              "Goroutines are lightweight threads managed by the Go runtime..."
            ]
          }
        },
        "type": "object",
        "required": [
          "answer_text"
        ],
        "title": "QuizAnswerCreate",
        "description": "Submit a free-text answer to a quiz question."
      },
      "QuizAnswerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Answer ID"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id",
            "description": "The question this answers"
          },
          "answer_text": {
            "type": "string",
            "title": "Answer Text",
            "description": "The submitted answer text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the answer was submitted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "answer_text",
          "created_at"
        ],
        "title": "QuizAnswerResponse",
        "description": "A recorded answer to a quiz question."
      },
...
      "QuizSessionDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Session ID"
          },
          "status": {
            "$ref": "#/components/schemas/QuizSessionStatus",
            "description": "Session status: pending, completed, or evaluated"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "description": "Number of questions in this session"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "When all answers were submitted (null if still pending)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the session was generated"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            },
            "type": "array",
            "title": "Questions",
            "description": "All questions in this session, ordered by index",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "question_count",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "QuizSessionDetailResponse",
        "description": "Quiz session with all questions, submitted answers, and LLM evaluations."
      },
      "QuizSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Session ID"
          },
          "status": {
            "$ref": "#/components/schemas/QuizSessionStatus",
            "description": "Session status: pending, completed, or evaluated"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "description": "Number of questions in this session"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "When all answers were submitted (null if still pending)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the session was generated"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "question_count",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "QuizSessionResponse",
        "description": "Quiz session summary."
      },
      "QuizSessionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "evaluated"
        ],
        "title": "QuizSessionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyProjectDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Project ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Project title",
            "examples": [
              "HTTP Rate Limiter"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Project overview and goals"
          },
          "difficulty_level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty Level",
            "description": "Difficulty level (1–5), calibrated to the user's profile"
          },
          "project_path": {
            "type": "string",
            "title": "Project Path",
            "description": "Filesystem path to the generated project codebase",
            "examples": [
              "workspace/projects/2026-w15-rate-limiter"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Project status: issued, submitted, or evaluated"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At",
            "description": "When the project was issued"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At",
            "description": "When the user submitted the project (null if not yet submitted)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/ProjectTaskResponse"
            },
            "type": "array",
            "title": "Tasks",
            "description": "Ordered list of tasks to complete",
            "default": []
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectEvaluationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "AI evaluation of the submission (null if not yet evaluated)"
          },
          "metadata_": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional project metadata (JSONB)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty_level",
          "project_path",
          "status",
          "issued_at",
          "submitted_at",
          "created_at",
          "updated_at"
        ],
        "title": "WeeklyProjectDetailResponse",
        "description": "Project with tasks, metadata, and optional AI evaluation."
      },
      "WeeklyProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Project ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Project title",
            "examples": [
              "HTTP Rate Limiter"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Project overview and goals"
          },
          "difficulty_level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty Level",
            "description": "Difficulty level (1–5), calibrated to the user's profile"
          },
          "project_path": {
            "type": "string",
            "title": "Project Path",
            "description": "Filesystem path to the generated project codebase",
            "examples": [
              "workspace/projects/2026-w15-rate-limiter"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Project status: issued, submitted, or evaluated"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At",
            "description": "When the project was issued"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At",
            "description": "When the user submitted the project (null if not yet submitted)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty_level",
          "project_path",
          "status",
          "issued_at",
          "submitted_at",
          "created_at",
          "updated_at"
        ],
        "title": "WeeklyProjectResponse",
        "description": "Weekly Go project summary."
      }
    }
  },
...
}